move wizard create logic
