+++ src/views/catalog/utils/tests/vm-produce.test.ts
@@ -2,6 +2,12 @@ import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 
 import { ensurePath } from '../WizardVMContext/utils/vm-produce';
 
+jest.mock('@openshift-console/dynamic-plugin-sdk', () => {
+  return {
+    k8sCreate: jest.fn().mockResolvedValue({}),
+  };
+});
+
 describe('Test ensurePath', () => {
   describe('Only one path', () => {
     it('Non existing path', () => {
